refactor: AgreementContentOptions
Refactored agreement content options to have value, id pairs stored within a constant and now use intl.format message to translate id stings within components
