+++ src/constants/agreementContentOptions.js
@@ -1,0 +1,42 @@
+export default [
+  {
+    value: 'alternateNames',
+    id: 'ui-agreements.alternativeNames',
+  },
+  {
+    value: 'agreementContentTypes',
+    id: 'ui-agreements.agreements.contentTypes',
+  },
+  {
+    value: 'contacts',
+    id: 'ui-agreements.agreements.internalContacts',
+  },
+  {
+    value: 'orgs',
+    id: 'ui-agreements.agreements.organizations',
+  },
+  {
+    value: 'items',
+    id: 'ui-agreements.agreementLines',
+  },
+  {
+    value: 'linkedLicenses',
+    id: 'ui-agreements.duplicateAgreementModal.linkedLicenses',
+  },
+  {
+    value: 'externalLicenseDocs',
+    id: 'ui-agreements.duplicateAgreementModal.externalLicenses',
+  },
+  {
+    value: 'supplementaryDocs',
+    id: 'ui-agreements.duplicateAgreementModal.supplementaryDocuments',
+  },
+  {
+    value: 'usageDataProviders',
+    id: 'ui-agreements.duplicateAgreementModal.usageData',
+  },
+  {
+    value: 'tags',
+    id: 'ui-agreements.agreements.tags',
+  },
+];
