+++ src/hooks/index.js
@@ -9,4 +9,3 @@ export { default as useAgreementsRefdata } from './useAgreementsRefdata';
 export { default as useEresourcesEnabled } from './useEresourcesEnabled';
 export { default as useBasket } from './useBasket';
 export { default as useBasketStore } from './useBasketStore';
-export { default as useAgreementContentOptions } from './useAgreementContentOptions';
